+++ CHANGELOG.md
@@ -1,13 +1,16 @@
 master
 ======
 
+Version 0.9.2
+=============
+
 Added
 -----
 * any header attribute can be changed in the config file
 
 Fixed
 -----
-* `doContour` and `doCut` option changed to lowercase
+* `doContour` and `doCut` options changed to lowercase
 
 
 Version 0.9.1